agent/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-classic>=0.0.1
langchain-anthropic>=0.1.0
anthropic>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0
